requirements.txt
42cc-pystyle==0.0.15
aiohttp==0.16.3
amqp==1.4.7
anyjson==0.3.3
args==0.1.0
beautifulsoup4==4.3.2
billiard==3.3.0.20
celery==3.1.8
certifi==0.0.8
cffi==1.1.2
chardet==2.3.0
cryptography==0.9.3
cssselect==0.9.1
Django==1.8.4
django-braces==1.4.0
django-crispy-forms==1.4.0
django-dajax==0.9.2
django-dajaxice==0.7
django-endless-pagination==2.0
django-haystack==2.4.0
django-mptt==0.7.4
django-password-reset==0.7
django-pure-pagination==0.2.1
djangoajax==2.3.1
elasticsearch==1.6.0
flake8==2.5.0
google-api-python-client==1.4.0
grab==0.6.29
gspread==0.2.5
HTMLParser==0.0.2
httplib2==0.9.1
idna==2.0
jdcal==1.0
kombu==3.0.28
lxml==3.4.4
mccabe==0.3.1
mock==1.3.0
nose==1.3.7
numpy==1.9.2
oauth2client==1.4.11
openpyxl==2.2.3
pandas==0.16.1
pbr==1.8.1
peewee==2.6.3
pep8==1.5.7
py==1.4.30
pyasn1==0.1.7
pyasn1-modules==0.0.5
pycparser==2.14
pycrypto==2.6.1
pycurl==7.19.5.1
PyDrive==1.0.1
pyflakes==1.0.0
pymongo==3.0.2
pyOpenSSL==0.15.1
pysolr==3.3.2
pytest==2.8.3
python-dateutil==2.4.2
pytils==0.3
pytz==2015.4
PyYAML==3.11
queuelib==1.2.2
requests==2.7.0
rsa==3.1.4
selection==0.0.11
selenium==2.48.0
simplejson==3.7.2
six==1.10.0
South==1.0.2
tqdm==1.0
Twisted==15.3.0
uritemplate==0.6
urllib3==1.11
user-agent==0.1.0
w3lib==1.12.0
weblib==0.1.20
xlrd==0.9.3
xlwt==1.0.0
zope.interface==4.1.2
